E with a grave accent now behaves just like every other character
Not sure why this plugin explicitly un-escaped "è" from "%C3%A8" to "è" (while leaving every other non-ASCII character as-is).
If we un-escape it, the relocation works, but then Rhythmbox would store the file path with a weird '&#'-escape in rythmdb.xml. This would make Rhythmbox unable to find the file again when it restarts.